package organization and naming cleanup
