+++ src/main/java/com/app/food/team/foodapp/dto/menurecipedto/RecipeDto.java
@@ -1,20 +1,12 @@
-package com.app.food.team.foodapp.dto;
-import com.app.food.team.foodapp.model.Item;
-import com.app.food.team.foodapp.model.ItemIngredient;
-import com.app.food.team.foodapp.model.TagDietLabel;
-import com.app.food.team.foodapp.repository.ItemRepository;
+package com.app.food.team.foodapp.dto.menurecipedto;
+import com.app.food.team.foodapp.model.menurecipemodel.Item;
+import com.app.food.team.foodapp.model.menurecipemodel.ItemIngredient;
+import com.app.food.team.foodapp.model.menurecipemodel.TagDietLabel;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
-import lombok.experimental.SuperBuilder;
 import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-
-import java.time.LocalDateTime;
-import java.util.Map;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
