+++ src/main/java/com/app/food/team/foodapp/model/menurecipemodel/ItemIngredient.java
@@ -1,26 +1,29 @@
-package com.app.food.team.foodapp.model;
+package com.app.food.team.foodapp.model.menurecipemodel;
 
+import com.app.food.team.foodapp.model.AbstractEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+@Getter
+@Setter
 @Entity
 @Table(name = "ingredients")
 @NoArgsConstructor
-//@AllArgsConstructor
+@AllArgsConstructor
 public class ItemIngredient extends AbstractEntity {
 
     @ManyToOne(cascade = CascadeType.ALL)
 //    @JoinColumn(insertable = false,updatable = false)
-    private @Getter @Setter Item item;
-    private @Getter @Setter String text;
-    private @Getter @Setter int quantity;
-    private @Getter @Setter String measure;
-    private @Getter @Setter String food;
-    private @Getter @Setter int weight;
-    private @Getter @Setter String foodId;
+    private Item item;
+    private String text;
+    private int quantity;
+    private String measure;
+    private String food;
+    private int weight;
+    private String foodId;
 
     @Override
     public String toString() {
@@ -34,13 +37,3 @@ public class ItemIngredient extends AbstractEntity {
                 '}';
     }
 }
-
-
-//        {
-//        "text": "string",
-//        "quantity": 0,
-//        "measure": "string",
-//        "food": "string",
-//        "weight": 0,
-//        "foodId": "string"
-//        }
