+++ src/main/java/com/app/food/team/foodapp/model/menurecipemodel/Menu.java
@@ -1,5 +1,6 @@
-package com.app.food.team.foodapp.model;
+package com.app.food.team.foodapp.model.menurecipemodel;
 
+import com.app.food.team.foodapp.model.AbstractEntity;
 import jakarta.persistence.*;
 import lombok.*;
