+++ src/main/java/com/app/food/team/foodapp/model/menurecipemodel/RecipeApiWrapper.java
@@ -1,4 +1,4 @@
-package com.app.food.team.foodapp.model;
+package com.app.food.team.foodapp.model.menurecipemodel;
 
 public class RecipeApiWrapper {
     private Item recipe;
