+++ src/main/java/com/app/food/team/foodapp/repository/authrepository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.app.food.team.foodapp.repository;
+package com.app.food.team.foodapp.repository.authrepository;
 
 import com.app.food.team.foodapp.model.authmodel.User;
 import org.springframework.data.jpa.repository.JpaRepository;
