+++ src/main/java/com/app/food/team/foodapp/repository/menureciperepository/ItemIngredientRepository.java
@@ -1,6 +1,6 @@
-package com.app.food.team.foodapp.repository;
+package com.app.food.team.foodapp.repository.menureciperepository;
 
-import com.app.food.team.foodapp.model.ItemIngredient;
+import com.app.food.team.foodapp.model.menurecipemodel.ItemIngredient;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
