+++ src/main/java/com/app/food/team/foodapp/repository/menureciperepository/TagDietLabelsRepository.java
@@ -1,6 +1,6 @@
-package com.app.food.team.foodapp.repository;
+package com.app.food.team.foodapp.repository.menureciperepository;
 
-import com.app.food.team.foodapp.model.TagDietLabel;
+import com.app.food.team.foodapp.model.menurecipemodel.TagDietLabel;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
